Rework the header
I added a placeholder for accepted_retrospectives so that we can use all
`_pages/` mechanism to specify the page order instead of hacking 2 of
them in the html layout as it is done in ai-folio. The placeholder
should never appear in the actual website as
`accepted_retrospectives/index.html` will have precedence.

Also turned the ML retrospective header into a link to home and made it
stick on all pages.

The about page is now Team.